==> wip_notebooks/1_Initial_data_API.ipynb <==
import sys
from pathlib import Path

parent = str(Path().resolve().parent)
if parent not in sys.path:
    sys.path.insert(0, parent)

import pandas as pd
import re

from functions.loading_articles import get_music_articles

with open("../categories/categories_musicians.txt", "r") as file:
    categories_musicians = file.read().splitlines()

with open("../categories/categories_groups.txt", "r") as file:
    categories_groups = file.read().splitlines()

musicians = get_music_articles(categories_musicians)
music_groups = get_music_articles(categories_groups)

print(len(music_groups))
print(len(musicians))

df_groups = pd.DataFrame(music_groups)
df_groups = df_groups.drop(columns=["ns"])

df_musicians = pd.DataFrame(musicians)
df_musicians = df_musicians.drop(columns=["ns"])

# extra col for musicians
def remove_parentheses(name):
    return re.sub(r'\s*\(.*?\)', '', name).strip()

df_musicians["name"] = df_musicians["title"].apply(remove_parentheses)


#df_musicians.to_csv("data/musicians.csv", index=False)
#df_groups.to_csv("data/music_groups.csv", index=False)

df_groups

df_musicians
