==> wip_notebooks/2_fetching_GroupMembers.ipynb <==
import sys
from pathlib import Path

parent = str(Path().resolve().parent)
if parent not in sys.path:
    sys.path.insert(0, parent)

import pandas as pd

from functions.member_fetch_functions import members_to_df, webscrape_members

df_musicans = pd.read_csv("../data/wip/musicians.csv")
df_groups = pd.read_csv("../data/wip/music_groups.csv")

df_groups["member"] = df_groups["title"].apply(members_to_df)

empty_member_indices = df_groups[df_groups["member"].apply(lambda x: x == [])].index

df_groups.loc[empty_member_indices, "member"] = df_groups.loc[empty_member_indices, "title"].apply(webscrape_members)

# df_groups.to_csv("data/groups_wMembers.csv", index=False)
df_groups
