==> wip_notebooks/3_Cleaned_members.ipynb <==
import sys
from pathlib import Path

parent = str(Path().resolve().parent)
if parent not in sys.path:
    sys.path.insert(0, parent)

import pandas as pd
import ast
import re
from itertools import chain

df_groups = pd.read_csv("../data/wip/groups_wMembers.csv")
df_groups["member"] = df_groups["member"].apply(ast.literal_eval)

def remove_parentheses(name):
    return re.sub(r'\s*\(.*?\)', '', name).strip()

df_groups["cleaned_member"] = df_groups["member"].apply(
    lambda members: [remove_parentheses(member) for member in members])

df_groups["cleaned_member"] = df_groups["cleaned_member"].apply(
    lambda members: [member.split(" - ")[0].strip() for member in members 
                     if member and member[0].isupper() or member[:2] == ""]
)

df_groups["cleaned_member"] = df_groups["cleaned_member"].apply(
    lambda members: [member for member in members 
                     if member and member[-1].isalpha() or member[-2:] == ""]
)

def remove_quotes(name):
    return name.replace('"', '').replace("'", '').strip()

df_groups["cleaned_member"] = df_groups["cleaned_member"].apply(
    lambda members: [remove_quotes(member) for member in members])

musical_instruments = [
    'Sang', 'Sanger', 'Vokal', 'Guitar', 'El-guitar', 'Piano', 'Keys', 'Klaver', 'Drums', 'Trommer', 'Bas', 'Bass', 'Violin', 'Saxophone', 'Saxofon' 'Trumpet',
    'Flute', 'Fløjte', 'Cello', 'Clarinet', 'Harpe', 'Trombone', 'Synthesizer', 'Keyboard', 'Sang/Guitar'
]

df_groups["cleaned_member"] = df_groups["cleaned_member"].apply(
    lambda members: [member for member in members if member not in musical_instruments])

all_members = list(chain.from_iterable(df_groups["cleaned_member"]))

print(all_members)
len(all_members)

# df_groups.to_csv("data/groups_wMembers.csv", index=False)

df_groups
