==> wip_notebooks/3a_discogs_fetching_Groupmembers.ipynb <==
import sys
from pathlib import Path

parent = str(Path().resolve().parent)
if parent not in sys.path:
    sys.path.insert(0, parent)

#download python3-discogs-client
import discogs_client
import pandas as pd
import ast
import numpy as np

consumer_key = 'ykUodTXeeVzqTJkACRzX'
consumer_secret = 'fOWGGoeHRESxkDPVoPFQKJBaTtDrFFRn'

d = discogs_client.Client('CSS_memberfetch/1.0', consumer_key=consumer_key, consumer_secret=consumer_secret)

# Generate the authorize URL
request_token, request_token_secret, authorize_url = d.get_authorize_url()

authorize_url

# After authorization, Discogs will give you a verifier code
verifier = 'pEbSxZwxhr'
# Get the access token
access_token, access_token_secret = d.get_access_token(verifier)
#lad vær med at køre det her mere end 1 gang, så får man fejl.

d.set_token(access_token, access_token_secret)

def get_discogs_members(band_name):
    
    if band_name[-7:]==" (band)": # to avoid (band) ending for messing up search results
        band_name = band_name[:-7]
    
    results = d.search(band_name, type='artist') 
    if len(results)>0:
        i = 0
        while i < len(results):
            artist = results[i]
            if artist.name.lower() in band_name.lower():
                break
            i += 1
        
        if i != len(results):
            try:
                if hasattr(artist, 'members'):
                    return [member.name for member in artist.members]
            except discogs_client.exceptions.HTTPError as e:
                pass
    return []
    

# problem: nogle gange står navnene i profile afsnit

df_groups = pd.read_csv("../data/wip/groups_wMembers.csv")
df_discgroups = pd.read_csv("../data/wip/groups_wMembers2.csv")
df_groups = df_groups.drop("member", axis = 1)

### New fetch of the members where wiki-scraping only returned 1 member

df_discgroups["disc_members"] = [[] for _ in range(len(df_discgroups))]

df_discgroups

# getting indexs of relevant rows
missing_members = df_discgroups[df_discgroups["cleaned_member"].apply(lambda x: len(ast.literal_eval(x))<=1)].index
wiki_indices = df_groups[df_groups["cleaned_member"].apply(lambda x: x != "[]")].index
members_to_check = [i for i in missing_members if i in wiki_indices]

df_discgroups.loc[members_to_check, 'disc_members'] = df_discgroups.loc[members_to_check, 'title'].apply(get_discogs_members)

df_discgroups["all_members"] = df_discgroups.apply(lambda row: np.unique(ast.literal_eval(row["cleaned_member"]) + (row["disc_members"])), axis = 1)

df_discgroups["all_members"] = df_discgroups["all_members"].apply(lambda memberlist: [str(member) for member in memberlist])

df_discgroups

#df_discgroups.to_csv("data/groups_wMembers3.csv", index=False)
df_discgroups

### Original fetch of discogs members 
(the ones who didnt have any results from wikipedia)

empty_member_indices = df_groups[df_groups["cleaned_member"].apply(lambda x: x == "[]")].index
df_groups.loc[empty_member_indices, 'cleaned_member']= df_groups.loc[empty_member_indices, 'title'].apply(get_discogs_members)

# df_groups.to_csv("data/groups_wMembers2.csv", index=False)

# problem: nogle gange er der flere bands med samme navn på discogs (fx the fashion). 
# hvis ikke det danske band er det først, så er det svært at gøre noget ved det

### problem: discogs har et tal efter personer hvis der er flere med samme navn
# ikke et problemm internt, men hvis det sakl sammenlignes med wiki går det galt
# måske kan vi bare fjerne det og håbe der ikke er flere med samme navn der optræder.
