==> wip_notebooks/4_creating_musicians_wMembers.ipynb <==
import sys
from pathlib import Path

parent = str(Path().resolve().parent)
if parent not in sys.path:
    sys.path.insert(0, parent)

import numpy as np
import pandas as pd
import ast
from itertools import chain

df_groups = pd.read_csv("../data/wip/groups_wMembers.csv")
df_groups["cleaned_member"] = df_groups["cleaned_member"].apply(ast.literal_eval)

df_musicians = pd.read_csv("../data/wip/musicians.csv")

members = list(chain.from_iterable(df_groups["cleaned_member"]))
musicians = list(df_musicians["name"].values)

all_musicians = musicians + members
all_musicians = np.unique(np.asarray(all_musicians))

len(all_musicians)

df_allmusicians = pd.DataFrame(all_musicians)
df_allmusicians.columns = ['name']

# df_all.to_csv("data/musicians_wMembers.csv", index=False)

#error to fix: Some names are joined in one line (ex: "Jakob Lundorff, Christoffer Hein og Dragut Lugalzagosi")
