==> wip_notebooks/5_Creating_Network.ipynb <==
import sys
from pathlib import Path

parent = str(Path().resolve().parent)
if parent not in sys.path:
    sys.path.insert(0, parent)

import re
import numpy as np
import ast
import pandas as pd
import networkx as nx
import matplotlib.pyplot as plt
from collections import Counter
import pickle

df_groups = pd.read_csv("../data/wip/groups_wMembers3.csv")
#df_groups["all_members"] = df_groups["all_members"].apply(ast.literal_eval)

df_groups = df_groups[["title", "all_members"]]
df_groups

### Group Musician Graph

df_groups["edgelist"] = df_groups.apply(lambda row: [(row["title"], member) for member in ast.literal_eval(row["all_members"])], axis=1)
edgelist = [edge for sublist in df_groups["edgelist"] for edge in sublist]

df_groups

GroupMusicianG = nx.Graph()
GroupMusicianG.add_edges_from(edgelist)

len(GroupMusicianG.nodes), len(GroupMusicianG.edges)

list_of_connected_comp = list(nx.connected_components(GroupMusicianG))
print(len(list_of_connected_comp)) # number of connected components

sorted_list = sorted(list_of_connected_comp, key=len, reverse = True)
lengths = [len(c) for c in sorted_list]

### Group graph

all_groups = list([group for group in df_groups["title"].values if group in GroupMusicianG.nodes])
GroupsG = nx.projected_graph(GroupMusicianG, all_groups, multigraph=False)

# Save graph
#with open("../data/wip/groups_graph.pkl", "wb") as f:
#    pickle.dump(GroupsG, f)

### Musician Graph

df_groups["all_members"].values

all_musicians = list(np.unique([
    str(member)
    for row_list in df_groups["all_members"].values
    for member in ast.literal_eval(row_list)
    if member in GroupMusicianG.nodes and member != ''
]))
all_musicians = [str(member) for member in all_musicians]

MusiciansG = nx.projected_graph(GroupMusicianG, all_musicians, multigraph=False)

# Save graph
#with open("../data/wip/Musicians_graph.pkl", "wb") as f:
#    pickle.dump(MusiciansG, f)
