==> wip_notebooks/6_fetching_text.ipynb <==
import sys
from pathlib import Path

parent = str(Path().resolve().parent)
if parent not in sys.path:
    sys.path.insert(0, parent)

import numpy as np
import pandas as pd
import requests
from functions.loading_articles import get_article_content_by_id

df_groups = pd.read_csv("../data/wip/groups_wMembers2.csv")

df_groups["intro_text"] = df_groups["pageid"].apply(get_article_content_by_id)

df_groups["intro_text"].iloc[10]

#df_groups.to_csv("data/groups_wMembers_Text.csv", index=False)
